+++ sats-dna/src/main/kotlin/com/sats/dna/GreyScaleModifier.kt
@@ -1,0 +1,24 @@
+package com.sats.dna
+
+import androidx.compose.ui.draw.DrawModifier
+import androidx.compose.ui.geometry.Rect
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.graphics.ColorMatrix
+import androidx.compose.ui.graphics.Paint
+import androidx.compose.ui.graphics.drawscope.ContentDrawScope
+import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
+
+class GreyScaleModifier : DrawModifier {
+    override fun ContentDrawScope.draw() {
+        val saturationMatrix = ColorMatrix().apply { setToSaturation(0f) }
+        val paint = Paint().apply {
+            colorFilter = ColorFilter.colorMatrix(saturationMatrix)
+        }
+
+        drawIntoCanvas {
+            it.saveLayer(Rect(0f, 0f, size.width, size.height), paint)
+            drawContent()
+            it.restore()
+        }
+    }
+}
